+++ MarketOps.SystemExecutor/GPW/CommissionGPWBossa.cs
@@ -10,17 +10,17 @@ namespace MarketOps.SystemExecutor.GPW
     /// </summary>
     public class CommissionGPWBossa : ICommission
     {
-        private float CalculateStock(int volume, float price)
+        private float CalculateStock(float volume, float price)
         {
             return Math.Max(volume * price * 0.0038f, 5f).TruncateTo2ndPlace();
         }
 
-        private float CalculateIndexFuture(int volume)
+        private float CalculateIndexFuture(float volume)
         {
             return (volume * 9.9f).TruncateTo2ndPlace();
         }
 
-        public float Calculate(StockType stockType, int volume, float price)
+        public float Calculate(StockType stockType, float volume, float price)
         {
             switch (stockType)
             {
